Moved and renamed Object mapper factory
