+++ src/test/java/io/github/pmckeown/dependencytrack/ObjectMapperFactoryTest.java
@@ -1,4 +1,4 @@
-package io.github.pmckeown.dependencytrack.builders;
+package io.github.pmckeown.dependencytrack;
 
 import io.github.pmckeown.dependencytrack.project.Project;
 import kong.unirest.GenericType;
@@ -12,11 +12,11 @@ import java.util.List;
 
 import static org.junit.Assert.assertNotNull;
 
-public class ObjectMapperBuilderTest {
+public class ObjectMapperFactoryTest {
 
     @Test
     public void thatUnknownPropertiesAreIgnoredWhenDeserializingJson() throws Exception {
-        JacksonObjectMapper om = new JacksonObjectMapper(ObjectMapperBuilder.relaxedObjectMapper());
+        JacksonObjectMapper om = new JacksonObjectMapper(ObjectMapperFactory.relaxedObjectMapper());
 
         List<Project> projects = om.readValue(IOUtils.toString(FileUtils.openInputStream(
                 new File("src/test/resources/__files/api/v1/project/get-all-projects.json"))),
